Added Waiting room, Modified Export component and Modified Quiz Management

+++ client/src/app/components/bank-question/bank-question.component.ts
@@ -1,4 +1,6 @@
 import { Component, Input, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmationDialogComponent } from '@app/components/dialog-component/dialog-delete.component';
 import { QuestionService } from '@app/services/quiz/question.service';
 import { Question } from '@common/types';
 
@@ -11,7 +13,10 @@ export class BankQuestionComponent implements OnInit {
     @Input() visibility: string[];
     questionList: Question[] = [];
 
-    constructor(private questionService: QuestionService) {}
+    constructor(
+        private questionService: QuestionService,
+        public dialog: MatDialog,
+    ) {}
 
     ngOnInit(): void {
         this.getAllQuestions();
@@ -40,12 +45,23 @@ export class BankQuestionComponent implements OnInit {
     }
 
     deleteQuestion(question: Question): void {
-        // Approuvé par le prof, _id spécifié par MongoDB
-        // eslint-disable-next-line no-underscore-dangle
-        this.questionService.deleteQuestionById(question._id).subscribe({
-            next: () => {
-                this.questionList.splice(this.questionList.indexOf(question), 1);
+        const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+            width: '350px',
+            data: {
+                title: 'Confirmation de la suppression',
+                message: 'Êtes-vous sûr de vouloir supprimer cette question?',
             },
+        });
+        dialogRef.afterClosed().subscribe((confirm) => {
+            if (confirm) {
+                // eslint-disable-next-line no-underscore-dangle
+                this.questionService.deleteQuestionById(question._id).subscribe({
+                    next: () => {
+                        // eslint-disable-next-line no-underscore-dangle
+                        this.questionList = this.questionList.filter((quest) => question._id !== quest._id);
+                    },
+                });
+            }
         });
     }
